+++ src/kvsc/main.rs
@@ -174,14 +174,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 std::process::exit(0x0001);
             }
             // for value perform input validation
-            if _value_input == INPUT_CLI {
-                if !input_validation::validate_value(
+            if _value_input == INPUT_CLI
+                && !input_validation::validate_value(
                     sub_m.value_of("value").unwrap().to_string(),
                     false,
-                ) {
-                    eprintln!("Provided value invalid.");
-                    std::process::exit(0x0001);
-                }
+                )
+            {
+                eprintln!("Provided value invalid.");
+                std::process::exit(0x0001);
             }
 
             // Get values of arguments
