+++ src/tests/mod.rs
@@ -1,0 +1,8 @@
+/*
+*  tests module declarations
+*  SPDX-License-Identifier: MIT
+*  Copyright (C) 2020 Benjamin Schilling
+*/
+
+mod integration;
+pub mod test_utils;
